Fix home-page module. Remove trash
