+++ client/src/app/home-page/components/home-about/home-about.component.scss
