+++ client/src/app/home-page/components/home-main/home-main.component.scss
