+++ client/src/app/home-page/components/home-news/home-news.component.scss
