+++ client/src/app/home-page/components/news-card/news-card.component.ts
