+++ client/src/app/home-page/components/pet-slider/pet-slider.component.scss
@@ -3,8 +3,10 @@
 	display: flex;
 	flex-direction: column;
 	align-items: center;
+	margin-bottom: 60px;
 	 h1{
 		margin-bottom: 40px;
+		margin-top: 60px;
 		font-size: 60px;
 		font-style: bold;
 	 }
