+++ client/src/app/home-page/components/pet-slider/pet-slider.component.spec.ts
