+++ client/src/app/home-page/components/pet-wanted-slider/pet-wanted-slider.component.ts
