+++ client/src/app/home-page/home-page-routing.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+import { HomePageComponent } from './home-page.component'; 
+
+const routes: Routes = [
+  { 
+    path: '', 
+    component: HomePageComponent 
+  },
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forChild(routes)
+  ],
+  exports: [
+    RouterModule
+  ]
+})
+
+export class HomePageRoutingModule {}
