+++ client/src/app/home-page/home-page.component.html
@@ -1,8 +1,8 @@
 <app-home-main></app-home-main>
-<app-wrapper>
+<div class="wrapper">
   <app-pet-slider></app-pet-slider>
   <app-pet-wanted-slider></app-pet-wanted-slider>
   <app-home-donate></app-home-donate>
   <app-home-about></app-home-about>
   <app-home-news></app-home-news>
-</app-wrapper>
+</div>
