+++ client/src/app/page-layout/header/header.module.ts
@@ -2,8 +2,11 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HeaderRoutingModule } from './header-routing.module';
 
+
 @NgModule({
   declarations: [],
-  imports: [CommonModule, HeaderRoutingModule],
+  imports: [
+    CommonModule,
+    HeaderRoutingModule],
 })
 export class HeaderModule {}
