--- client/src/app/page-layout/home-page/home-page.module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { HomePageRoutingModule } from './home-page-routing.module';
-
-
-@NgModule({
-  declarations: [],
-  imports: [
-    CommonModule,
-    HomePageRoutingModule
-  ]
-})
-export class HomePageModule { }
