--- client/src/app/page-layout/wrapper/wrapper.component.html
@@ -1,3 +1,0 @@
-<div class="wrapper">
-  <ng-content></ng-content>
-</div>
