+++ client/src/app/pets/pages/animal-list-page/animal-list-page.component.html
@@ -2,10 +2,6 @@
   <div>
     <h1>{{title}}</h1>
     <app-animal-list-card></app-animal-list-card>
-    <p>
-      customers works!
-    </p>
-    
   </div>
 </div>
